refactor: move merchant account to domain models
