+++ crates/router/src/types/domain.rs
@@ -3,9 +3,14 @@ pub mod behaviour {
     pub use hyperswitch_domain_models::behaviour::{Conversion, ReverseConversion};
 }
 
+pub mod merchant_account {
+    pub use hyperswitch_domain_models::merchant_account::*;
+}
+
+pub use merchant_account::*;
+
 mod customer;
 mod event;
-mod merchant_account;
 mod merchant_connector_account;
 mod merchant_key_store {
     pub use hyperswitch_domain_models::merchant_key_store::MerchantKeyStore;
@@ -19,7 +24,6 @@ pub mod user_key_store;
 pub use address::*;
 pub use customer::*;
 pub use event::*;
-pub use merchant_account::*;
 pub use merchant_connector_account::*;
 pub use merchant_key_store::*;
 pub use payments::*;
